feat(sdk-lib-mpc): move shamir
TICKET: WP-400
