--- modules/sdk-core/src/account-lib/mpc/types.ts
@@ -1,4 +1,0 @@
-export interface SplitSecret {
-  shares: Record<number, bigint>;
-  v: Array<bigint>;
-}
