+++ modules/sdk-lib-mpc/src/curves/index.ts
@@ -2,3 +2,4 @@ export * from './types';
 export * from './secp256k1';
 export * from './bip32';
 export * from './hdTree';
+export * from './shamir';
